requirements.txt
# Data Science & ML
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.1.0
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.2

# Jupyter
jupyterlab==4.0.5
notebook==7.0.2
nbformat==5.9.2
nbclient==0.9.0
nbconvert==7.9.2

# AWS SDK (for evaluation runner inside container)
boto3==1.34.131
botocore==1.34.131

# Optional utils
pyarrow==14.0.2
joblib==1.3.2

python-dotenv==1.0.0
